// packages/contracts/src/generated/api.d.ts
export interface components {
    schemas: {
// ...
        KioskAccessoryResponse: {
            /** Format: uuid */
            id: string;
            name: string;
            /** Format: int64 */
            price: number | string;
            /** Format: int32 */
            stockQuantity: number | string;
            shopName: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}